Implement validating admission webhook for CR users.appuio.io
We use the "low-level" approach of registering our handler directly on
the controller manager's webhook server as we are able to access the
manager's Kubernetes client with that approach.

The client is not easily accessible when taking the "high-level"
approach of implementing the `Validator` interface on the type which we
want to validate.

Note that the service name and namespace in the
`ValidatingWebhookConfiguration` generated by controller-gen are not
configurable and need to adjusted during deployment.