Add default organization to users that are in exactly one organization
We want to add the usability feature "As a user, I want my default
organization to be set automatically when I create my first organization
or accept an invitation, so that I can get started right away and don't
have to worry about setup."

We simplify this to "every user who is in exactly 1 organization and
does not have a default configured should have their default
organization set to the one organization they are a member of". This
will automatically solve the use case above both for creating your first
org and for accepting your first invite.